cli-app/pyproject.toml
[project]
name = "ai-agent-cli"
version = "1.0.0"
description = "Beautiful CLI for AI agents powered by Ollama and Agno framework"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "AI Agent CLI", email = "agent@example.com"},
]
keywords = ["ai", "cli", "agent", "ollama", "chat", "assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "agno[aws]>=1.7.1",
    "anthropic>=0.57.1",
    "groq>=0.29.0",
    "ollama>=0.5.1",
    "python-dotenv>=1.1.1",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.41",
]

[project.scripts]
my-cli = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
